fix: split NSPopoverDelegate extension into PopoverController+Delegate.swift to satisfy file_length lint rule
